Major refactoring to reduce memory consumption.
Removed support for rereporting attributes when no default response was received and simplified where we send reports to. Instead of storing the address everywhere, we now only differentiate between sending attribute reports to the coordinator (and then we do store the endpoint), or a broadcast.

These changes allow the Sensors sketch (5 attributes with reporting configured) to run in 2 KB on a Arduino Pro Mini.